refund tests, init tests, memory tests passing

+++ src/Nevermind/Nevermind.Core/RefundOf.cs
@@ -1,10 +1,8 @@
-using System.Numerics;
-
 namespace Nevermind.Core
 {
     public static class RefundOf
     {
-        public static BigInteger SClear = 15000;
-        public static BigInteger Destroy = 24000;
+        public static ulong SClear = 15000;
+        public static ulong Destroy = 24000;
     }
 }
